Split firebase_storage plugin into multiple files. (#550)
